+++ projects/igniteui-angular/src/lib/directives/for-of/horizontal.virtual.helper.component.ts
@@ -8,7 +8,7 @@ import { Component, ElementRef, HostBinding, Input, ViewChild, ViewContainerRef 
     template: '<div #horizontal_container class="igx-vhelper__placeholder-content" [style.width.px]="width"></div>'
 })
 export class HVirtualHelperComponent {
-    @ViewChild('horizontal_container', { read: ViewContainerRef }) public _vcr;
+    @ViewChild('horizontal_container', { read: ViewContainerRef, static: true }) public _vcr;
     @Input() public width: number;
     @HostBinding('class')
     public cssClasses = 'igx-vhelper--horizontal';